pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "unrest"
version = "0.1.0"
description = "Lean and mean web framework for Python / Postgres"
authors = ["Working Proof <hello@workingproof.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
starlette = "^0.38.5"
uvicorn = "^0.30.6"
asyncpg = "^0.29.0"
httpx = "^0.27.2"
pydantic = "^2.9.1"
python-dotenv = "^1.0.1"
python-json-logger = "^2.0.7"
bcrypt = "^4.2.0"
taskiq = {extras = ["reload"], version = "^0.11.7"}
taskiq-redis = "^1.0.0"
asyncclick = "^8.1.7.2"
click = "^8.1.7"
jinja2 = "^3.1.4"
python-multipart = "^0.0.17"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
ruff = "^0.6.4"
pytest-asyncio = "^0.24.0"
asgi-lifespan = "^2.1.0"
pre-commit = "^3.8.0"
mkdocs-material = "^9.5.34"

[tool.ruff]
show-fixes = true
line-length = 150

[tool.ruff.lint]
extend-select = ["I"]

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "session"

[tool.poetry.scripts]
unrest = "unrest.cli:main"


